+++ pkg/pve/pve_node_firewall.go
@@ -1,9 +1,12 @@
 package pve
 
 import (
+	"fmt"
 	"net/http"
 	"path"
 	"strconv"
+
+	"github.com/google/uuid"
 )
 
 type PVENodeFirewallService struct {
@@ -54,6 +57,39 @@ func (s *PVENodeFirewallService) GetRulesByPos(node string, pos int) (GetNodeFir
 	err := s.api.client.sendReq(method, path, nil, res)
 
 	return *res, err
+}
+
+type CreateNodeFirewallRuleRequest struct {
+	Action          string           `in:"nonzero;form=action"`      // Rule action ('ACCEPT', 'DROP', 'REJECT') or security group name. Format: [A-Za-z][A-Za-z0-9\-\_]+.
+	Node            string           `in:"nonzero;path=node"`        // The cluster node name.
+	Type            string           `in:"nonzero;form=type"`        // Rule type. Values: in | out | forward | group.
+	Comment         string           `in:"omitempty;form=comment"`   // Optional. Descriptive comment.
+	Destination     string           `in:"omitempty;form=dest"`      // Optional. Restrict packet destination address. This can refer to a single IP address, an IP set ('+ipsetname') or an IP alias definition. You can also specify an address range like '20.34.101.207-201.3.9.99', or a list of IP addresses and networks (entries are separated by comma). Please do not mix IPv4 and IPv6 addresses inside such lists.
+	Digest          string           `in:"omitempty;form=digest"`    // Optional. Prevent changes if current configuration file has a different digest. This can be used to prevent concurrent modifications.
+	DestinationPort int              `in:"omitempty;form=dport"`     // Optional. Restrict TCP/UDP destination port. You can use service names or simple numbers (0-65535), as defined in '/etc/services'. Port ranges can be specified with '\d+:\d+', for example '80:85', and you can use comma separated list to match several ports or ranges.
+	Enable          int              `in:"omitempty;form=enable"`    // Optional. Flag to enable/disable a rule. Values: (0 - N).
+	ICMPType        string           `in:"omitempty;form=icmp-type"` // Optional. Specify icmp-type. Only valid if proto equals 'icmp' or 'icmpv6'/'ipv6-icmp'.
+	Interface       string           `in:"omitempty;form=iface"`     // Optional. Network interface name. You have to use network configuration key names for VMs and containers ('net\d+'). Host related rules can use arbitrary strings.
+	LogLevel        FirewallLogLevel `in:"omitempty;form=log"`       // Optional. Log level for firewall rule.
+	Macro           string           `in:"omitempty;form=macro"`     // Optional. Use predefined standard macro.
+	Pos             int              `in:"omitempty;form=pos"`       // Optional. Update rule at position <pos>.
+	Proto           string           `in:"omitempty;form=proto"`     // Optional. IP protocol. You can use protocol names ('tcp'/'udp') or simple numbers, as defined in '/etc/protocols'.
+	Source          string           `in:"omitempty;form=source"`    // Optional. Restrict packet source address. This can refer to a single IP address, an IP set ('+ipsetname') or an IP alias definition. You can also specify an address range like '20.34.101.207-201.3.9.99', or a list of IP addresses and networks (entries are separated by comma). Please do not mix IPv4 and IPv6 addresses inside such lists.
+	Sport           string           `in:"omitempty;form=sport"`     // Optional. Restrict TCP/UDP source port. You can use service names or simple numbers (0-65535), as defined in '/etc/services'. Port ranges can be specified with '\d+:\d+', for example '80:85', and you can use comma separated list to match several ports or ranges.
+}
+
+// NewRule creates a new node firewall rule. It adds metadata within the proxmox comment field with the format [id=uuid].
+//
+// POST /nodes/:node/firewall/rules requires the "Sys.Modify" permission.
+func (s *PVENodeFirewallService) NewRule(req CreateNodeFirewallRuleRequest) (string, error) {
+	method := http.MethodPost
+	path := "/nodes/{node}/firewall/rules"
+	uuid := uuid.New().String()
+	req.Comment = fmt.Sprintf("[id=%s] %s", uuid, req.Comment)
+	if err := s.api.client.sendReq2(method, path, &req, nil); err != nil {
+		return "", err
+	}
+	return uuid, nil
 }
 
 // ReadLog Retrieves node's firewall log entries.