+++ CHANGELOG.md
@@ -6,12 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
-
-## [v0.5.1]
 ### PVE API wrapper
 #### Fixed
 - pve api wrapper host option is now `--host`.
 - pve api wrapper port option is now `--port`.
+
+### PVE API client
+#### Added
+- `POST /nodes/{node}/firewall/rules` implementation as `PVE.Node.Firewall.NewRule`.
 
 ## [v0.5.0]
 ### PVE API wrapper